add helpers to get started with fuse faster
An init container has been added to docker-compose.yml
to simplify initial configuration. Migration, Seeds and Metadata.

Additionally, the init container has a helper script to simplify
adding a new repo for quicker testing.

If the first argument is `add-repo` the provided repo will be
inserted and the appropriate sync types will be added.

A default grafana setup has been included as well with a simple
example dashboard with a default postgres datasource.