feat: add pg function for detecting errors in a sync job
this adds a pg function that's meant to be consumed as a computed column in Graphile (https://www.graphile.org/postgraphile/computed-columns/). The function `mergestat.repo_sync_queue_has_error` (which should appear in the GraphQL query just as `hasError`) looks up whether a repo sync job has an error log, and returns a boolean `true` or `false.