Sign all container images with sigstore's cosign
This uses cosign to sign the graphql, UI and worker images. Cosign is
configured to use keyless signing and will use a GitHub OIDC token
provided to the workflow in order to request a certificate and do the
signature.

Signatures are posted to a public transparency log (rekor.sigstore.dev)
and the signature itself is pushed to the same repository as the images
themselves.

To verify, do:

    COSIGN_EXPERIMENTAL=1 cosign verify <image name>

Signed-off-by: Juan Antonio Osorio <juan.osoriorobles@eu.equinix.com>