fix: remove `commit_stats_pkey` PK for now
This is to address new deployments that encounter this migration. They _may_ hit a failure on duplicate key constraints because the columns here are not guaranteed to be unique (`file_path` alone is not unique enough within a commit, we saw an error with a submodule replacing a file).

By removing this migration, we protect future deployments from running this on against existing data.