fix: remove `.env-worker` requirement for `docker-compose`
For now, just use inline env vars in the `docker-compose.yaml` file. We can configure a `make` target for a more complex setup that uses `.env-worker` later if we need it.

This allows the app to be brought up on a "fresh" clone of the repo with just `docker-compose up`. Closes #418