fix: update the `docker-compose` file to address startup timing issues
we were seeing problems when running `docker-compose up`, in particular the `graphql` service would start before the migrations in the `worker` service were complete. This would lead to an invalid GraphQL API schema being served (no entities) and errors displaying in the UI. Not a great UX. The solution was to just rerun `docker-compose up` and hope the timing issue didn't occur again.

This addresses that (it should anyways) by adding health checks to some services and `condition: service_healthy` in the `depends_on` config in some spots.

Note it also brings back the `DEBUG=1` env var, which is what enables the `/metrics` endpoint in the `worker`, which is what we're using a healthcheck now - this is probably worth improving.

It also bumps the default `CONCURRENCY` to `5` for `docker-compose` runs.