feat: add a `disableReadOnly` flag to the `execSQL` graphql query
This will allow both readonly and non-readonly queries to be run in the UI, based on what the API request sends. This is distinct from user permissions, and is meant more to be a "break glass" type of param to protect users running queries by defaulting to readonly (and explicitly requiring them to opt-in to non-readonly query access).