feat: add `includeArchivedRepos ` auto import config flag
This changes the default behavior of GitHub repo auto imports to _skip_ archived repos. The prior behavior (including archived repos) can be configured by setting the `includeArchivedRepos` field to `true`.

There is currently no place in the UI where this is changeable,  so it will have to be done via direct SQL for now.